[korean] adapt to LaTeX 2021/05/01 pre-release-2
```
\documentclass{article}
\usepackage{polyglossia}
\setmainlanguage[variant=classic]{korean}
\newfontfamily\hangulfont{UnBatang.ttf}[Script=Hangul]
\begin{document}

가^^^^3002\par
나다

\end{document}
```
Upon running this file with `xelatex-dev` included in TLpretest
(LaTeX 2021/05/01 pre-release-2), we get a result in which `\par` is
disappeared.  The reason is that LaTeX 2021/05/01 redefined `\par`,
such that `\par` is called while gloss-korean.ldf has not yet finished
the making of hboxes.  The commit aims to adapt to this change of
LaTeX format. Now hboxes are completed before `\par` is called.

(Another solution is to add `\relax` to the beginning of the definiton
of `\par`. This one is much simpler, But is beyond our control.)