Fix (re)setting of scripts in inline language commands
As it was, \resetfontlatin has been called if the language to be closed
was non-Latin script, no matter if the embedding (now active) language
actually is of Latin script.

No we check the latter instead.

This bug was visible in example-arabic, where the tt font in the Farsi
footnote was wrong and the text invisible.