Avoid superfluous writing of language switches to aux file (#593)
It turns out this is not only non-sensical, but potentially harmful.

For instance the title of our manual was in ancient Greek after a second
latex pass as the latest language switch to Greek used that variant, and
after the aux file has been read at document begin, this setting stuck
for the optionless \textgreek we use in the title.