re-implementing \charifavailable using \iffontchar (fix #613)
The primitive \iffontchar is available in e-TeX and so is available in
both XeTeX and LuaTeX. See section 3.12 in e-TeX's manual for details.

Replacing \char with \Uchar will also make the command expandable.

The reasson for using \exp_arg:Nno to expand \Uchar"#1 is that it
prevents the look ahead problem described in ticket #613.